AppInsights metrics exporter simplifications + perf enhancements.
Two things:
- The sleep interval calculation was wrong because the stopwatch was never reset. sw.Elapsed would continue growing. I just removed this code.
- Thread.Sleep was taking a couple percent of my CPU while running a small load test so I removed it in favor of Task.Delay.
