+++ src/OpenTelemetry.Exporter.ApplicationInsights/ApplicationInsightsExporter.cs
@@ -17,11 +17,15 @@
 namespace OpenTelemetry.Exporter.ApplicationInsights
 {
     using System;
+    using System.Diagnostics;
     using System.Threading;
     using System.Threading.Tasks;
+    using Microsoft.ApplicationInsights;
+    using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.Extensibility;
     using OpenTelemetry.Exporter.ApplicationInsights.Implementation;
     using OpenTelemetry.Stats;
+    using OpenTelemetry.Stats.Aggregations;
     using OpenTelemetry.Trace.Export;
 
     /// <summary>
@@ -33,9 +37,13 @@ namespace OpenTelemetry.Exporter.ApplicationInsights
 
         private readonly TelemetryConfiguration telemetryConfiguration;
 
+        private readonly TelemetryClient telemetryClient;
+
         private readonly IViewManager viewManager;
 
         private readonly ISpanExporter exporter;
+
+        private readonly TimeSpan exportMetricsInterval;
 
         private readonly object lck = new object();
 
@@ -43,7 +51,7 @@ namespace OpenTelemetry.Exporter.ApplicationInsights
 
         private CancellationTokenSource tokenSource;
 
-        private Task workerThread;
+        private Task exportMetricsTask;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ApplicationInsightsExporter"/> class.
@@ -52,11 +60,18 @@ namespace OpenTelemetry.Exporter.ApplicationInsights
         /// <param name="exporter">Exporter to get traces and metrics from.</param>
         /// <param name="viewManager">View manager to get stats from.</param>
         /// <param name="telemetryConfiguration">Telemetry configuration to use to report telemetry.</param>
-        public ApplicationInsightsExporter(ISpanExporter exporter, IViewManager viewManager, TelemetryConfiguration telemetryConfiguration)
+        /// <param name="exportMetricsInterval">The export metrics interval. Defaults to 1 minute.</param>
+        public ApplicationInsightsExporter(
+            ISpanExporter exporter,
+            IViewManager viewManager,
+            TelemetryConfiguration telemetryConfiguration,
+            TimeSpan? exportMetricsInterval = null)
         {
             this.exporter = exporter;
             this.viewManager = viewManager;
             this.telemetryConfiguration = telemetryConfiguration;
+            this.telemetryClient = new TelemetryClient(telemetryConfiguration);
+            this.exportMetricsInterval = exportMetricsInterval ?? TimeSpan.FromMinutes(1);
         }
 
         /// <summary>
@@ -77,10 +92,7 @@ namespace OpenTelemetry.Exporter.ApplicationInsights
 
                 this.tokenSource = new CancellationTokenSource();
 
-                var token = this.tokenSource.Token;
-
-                var metricsExporter = new MetricsExporterThread(this.telemetryConfiguration, this.viewManager, token, TimeSpan.FromMinutes(1));
-                this.workerThread = Task.Factory.StartNew((Action)metricsExporter.WorkerThread, TaskCreationOptions.LongRunning);
+                this.exportMetricsTask = this.ExportMetricsAsync(this.exportMetricsInterval, this.tokenSource.Token);
             }
         }
 
@@ -98,10 +110,161 @@ namespace OpenTelemetry.Exporter.ApplicationInsights
 
                 this.exporter.UnregisterHandler(TraceExporterName);
                 this.tokenSource.Cancel();
-                this.workerThread.Wait();
+                this.exportMetricsTask.Wait();
                 this.tokenSource = null;
 
                 this.handler = null;
+            }
+        }
+
+        /// <summary>
+        /// Starts the metrics export task.
+        /// </summary>
+        /// <param name="exportInterval">The export interval.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>A Task that when complete signals the end of the continous Export.</returns>
+        private async Task ExportMetricsAsync(TimeSpan exportInterval, CancellationToken cancellationToken)
+        {
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                try
+                {
+                    await Task.Delay(exportInterval, cancellationToken).ConfigureAwait(false);
+
+                    try
+                    {
+                        this.Export();
+                    }
+                    catch (Exception ex)
+                    {
+                        // TODO Log to something useful.
+                        Debug.WriteLine(ex);
+                    }
+                }
+                catch (TaskCanceledException)
+                {
+                    break;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Exports this instance.
+        /// </summary>
+        private void Export()
+        {
+            foreach (var view in this.viewManager.AllExportedViews)
+            {
+                var data = this.viewManager.GetView(view.Name);
+
+                foreach (var value in data.AggregationMap)
+                {
+                    var metricTelemetry = new MetricTelemetry
+                    {
+                        Name = data.View.Name.AsString,
+                    };
+
+                    for (var i = 0; i < value.Key.Values.Count; i++)
+                    {
+                        var name = data.View.Columns[i].Name;
+                        var val = value.Key.Values[i].AsString;
+                        metricTelemetry.Properties.Add(name, val);
+                    }
+
+                    // Now those properties needs to be populated.
+                    //
+                    // metricTelemetry.Sum
+                    // metricTelemetry.Count
+                    // metricTelemetry.Max
+                    // metricTelemetry.Min
+                    // metricTelemetry.StandardDeviation
+                    //
+                    // See data model for clarification on the meaning of those fields.
+                    // https://docs.microsoft.com/azure/application-insights/application-insights-data-model-metric-telemetry
+
+                    value.Value.Match<object>(
+                        (combined) =>
+                        {
+                            if (combined is ISumDataDouble sum)
+                            {
+                                metricTelemetry.Sum = sum.Sum;
+                            }
+
+                            return null;
+                        },
+                        (combined) =>
+                        {
+                            if (combined is ISumDataLong sum)
+                            {
+                                metricTelemetry.Sum = sum.Sum;
+                            }
+
+                            return null;
+                        },
+                        (combined) =>
+                        {
+                            if (combined is ICountData count)
+                            {
+                                metricTelemetry.Sum = count.Count;
+                            }
+
+                            return null;
+                        },
+                        (combined) =>
+                        {
+                            if (combined is IMeanData mean)
+                            {
+                                metricTelemetry.Sum = mean.Mean * mean.Count;
+                                metricTelemetry.Count = (int)mean.Count;
+                                metricTelemetry.Max = mean.Max;
+                                metricTelemetry.Min = mean.Min;
+                            }
+
+                            return null;
+                        },
+                        (combined) =>
+                        {
+                            if (combined is IDistributionData dist)
+                            {
+                                metricTelemetry.Sum = dist.Mean * dist.Count;
+                                metricTelemetry.Count = (int)dist.Count;
+                                metricTelemetry.Min = dist.Min;
+                                metricTelemetry.Max = dist.Max;
+                                metricTelemetry.StandardDeviation = dist.SumOfSquaredDeviations;
+                            }
+
+                            return null;
+                        },
+                        (combined) =>
+                        {
+                            if (combined is ILastValueDataDouble lastValue)
+                            {
+                                metricTelemetry.Sum = lastValue.LastValue;
+                            }
+
+                            return null;
+                        },
+                        (combined) =>
+                        {
+                            if (combined is ILastValueDataLong lastValue)
+                            {
+                                metricTelemetry.Sum = lastValue.LastValue;
+                            }
+
+                            return null;
+                        },
+                        (combined) =>
+                        {
+                            if (combined is IAggregationData aggregationData)
+                            {
+                                // TODO: report an error
+                            }
+
+                            return null;
+                        });
+
+                    this.telemetryClient.TrackMetric(metricTelemetry);
+                }
             }
         }
     }
